Remove unnecessary file diffs
